fix: rename client_go_adaper to client_go_adapter
